pyproject.toml
[tool.poetry]
name = "finacial-intelligence"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
torch = "^2.3.1"
scikit-learn = "^1.5.0"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
